requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
anthropic==0.26.1
pdfplumber==0.11.0
python-docx==1.1.2
python-multipart==0.0.9
